requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.21.0
python-dotenv==1.0.1
loguru==0.7.2
pyyaml==6.0.1
pandas==2.2.2
tenacity==8.2.3
flask==2.3.3
